// sicurezza_alimentare_e_veterinaria/client/src/app/menu/types.ts
/*
Copyright (C) AGPL-3.0  

This program is free software: you can redistribute it and/or modify it under the terms of the GNU Affero General Public License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version.

This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Affero General Public License for more details.

You should have received a copy of the GNU Affero General Public License along with this program. If not, see <https://www.gnu.org/licenses/>.*/
export interface MenuItem {
  label?: string;
  icon?: string;
  link?: string;
  action?: Function;
  disabled?: boolean;
  lev?: number;
  ord?: number;
  modality?: string;
  hidden?: boolean;
}

export interface Menu {
  menuItems: MenuItem[];
}

export interface ToggleableMenu {
  openMenu(): void;
  closeMenu(): void;
}